Remove final for this class
This allows to extend this class and override some behaviours. My idea is to create class YankExtended that extends Yank and throws database errors up. Why? I do understand why you hide exceptions from code. But this stops anyone of creating more complex microservices, because if the database is down, I need to know it is down so I can put circuit breakers on the code. So to hide exceptions in all use cases in definitely the best solution.